pyproject.toml
[tool.poetry]
name = "nlip-with-agent-frameworks"
version = "1.0.0"
description = "NLIP integration demo with popular AI agent frameworks"
readme = "README.md"
package-mode = false
packages = [{include = "demo"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.12"
uvicorn = "^0.24.0"
httpx = "^0.25.2"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"

# NLIP dependencies
nlip-sdk = "^0.1.2"
nlip-server = {path = "./nlip_server", develop = true}

# LangChain dependencies
langchain = "^0.1.0"
langchain-openai = "^0.1.0"
langchain-core = "^0.1.0"

# LlamaIndex dependencies
setuptools = "^80.9.0"
llama-index = "0.12.45"
llama-index-core = "0.12.45"
llama-index-llms-openai-like = "0.4.0"
llama-index-workflows = "1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
start-inter-agent = "scripts.start_inter_agent:main"
start-standalone = "scripts.start_standalone:main"
test-demo = "scripts.test_demo:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
